refactor: pools and txs
